+++ app/Api/Models/UsersModel.php
@@ -1,0 +1,57 @@
+<?php
+
+namespace App\Api\Models;
+
+use App\Api\Core\Traits\UuidTrait;
+use Illuminate\Contracts\Auth\MustVerifyEmail;
+use Illuminate\Database\Eloquent\Factories\HasFactory;
+use Illuminate\Foundation\Auth\User as Authenticatable;
+use Illuminate\Notifications\Notifiable;
+use Laravel\Sanctum\HasApiTokens;
+
+class UsersModel extends Authenticatable
+{
+    use HasApiTokens, UuidTrait;
+
+    /**
+     * @var string
+     */
+    protected $table = 'users';
+
+    /**
+     * The attributes that should be hidden for serialization.
+     *
+     * @var array<int, string>
+     */
+    protected $hidden = [
+        'id',
+        'password',
+        'remember_token',
+    ];
+
+    /**
+     * @var array<string>
+     */
+    protected $dates = [
+        'created_at',
+        'updated_at'
+    ];
+
+    /**
+     * The attributes that are mass assignable.
+     *
+     * @var array<int, string>
+     */
+    protected $fillable = [
+        'name',
+        'surname',
+        'email',
+        'password',
+    ];
+
+    public function accountType()
+    {
+        return $this->hasOneThrough(AccountTypesModel::class, UserAccountsModel::class, 'user_id', 'id');
+
+    }
+}